add permission

+++ backend/src/permissions/permissions.service.ts
@@ -8,16 +8,35 @@ import { In, Repository } from 'typeorm';
 import { Permission } from './entities/permission.entity';
 import { CreatePermissionDto } from './dto/create-permission.dto';
 import { UpdatePermissionDto } from './dto/update-permission.dto';
+import { RolePermission } from 'src/roles/entities/role-permission.entity';
+import { MenuPermission } from 'src/menu/entities/menu-permission.entity';
 
 @Injectable()
 export class PermissionsService {
   constructor(
     @InjectRepository(Permission)
     private readonly permRepo: Repository<Permission>,
+    @InjectRepository(RolePermission)
+    private readonly rolePermRepo: Repository<RolePermission>,
+    @InjectRepository(MenuPermission)
+    private readonly menuPermRepo: Repository<MenuPermission>,
   ) {}
 
   async findAll(): Promise<Permission[]> {
     return this.permRepo.find({ order: { id: 'ASC' } });
+  }
+
+  async findBySearch(search: string): Promise<Permission[]> {
+    return await this.permRepo
+      .createQueryBuilder('perm')
+      .where('LOWER( perm.permission_code) LIKE LOWER(:search)', {
+        search: `%${search}%`,
+      })
+      .orWhere('LOWER(perm.permission_name) LIKE LOWER(:search)', {
+        search: `%${search}%`,
+      })
+      .orderBy('perm.id', 'ASC')
+      .getMany();
   }
 
   async findOne(id: number): Promise<Permission> {
@@ -31,19 +50,51 @@ export class PermissionsService {
       where: { permission_code: dto.permission_code },
     });
     if (existing) throw new ConflictException('Permission code đã tồn tại');
-    return this.permRepo.save(this.permRepo.create(dto));
+    return this.permRepo.save(dto);
   }
-
+  //update permission nếu có thay đổi permission_code thì phải kiểm tra trùng với permission khác chưa, nếu trùng thì throw lỗi
+  // trả về Permission sau khi update
   async update(id: number, dto: UpdatePermissionDto): Promise<Permission> {
-    await this.findOne(id);
-    await this.permRepo.update(id, dto);
-    return this.findOne(id);
+    const perm = await this.findOne(id);
+    if (!perm) {
+      throw new NotFoundException('Không tìm thấy permission');
+    }
+    if (dto.permission_code && dto.permission_code !== perm.permission_code) {
+      const existing = await this.permRepo.findOne({
+        where: { permission_code: dto.permission_code },
+      });
+      if (existing) {
+        throw new ConflictException('Permission code đã tồn tại');
+      }
+    }
+    Object.assign(perm, dto);
+    return this.permRepo.save(perm);
   }
 
   async remove(id: number): Promise<{ message: string }> {
     const perm = await this.findOne(id);
+    if (!perm) {
+      throw new NotFoundException('Không tìm thấy permission');
+    }
+    // xóa permission khỏi tất cả role trước khi xóa permission
+    await this.rolePermRepo.delete({ permission_id: id });
+    //xóa permission khỏi tất cả menu trước khi xóa permission
+    await this.menuPermRepo.delete({ permission_id: id });
     await this.permRepo.remove(perm);
     return { message: 'Xoá permission thành công' };
+  }
+
+  async deleteMultiple(ids: number[]): Promise<{ message: string }> {
+    try {
+      // xóa permission khỏi tất cả role trước khi xóa permission
+      await this.rolePermRepo.delete({ permission_id: In(ids) });
+      //xóa permission khỏi tất cả menu trước khi xóa permission
+      await this.menuPermRepo.delete({ permission_id: In(ids) });
+      await this.permRepo.delete({ id: In(ids) });
+      return { message: 'Xoá permissions thành công' };
+    } catch (error) {
+      throw new NotFoundException('Lỗi :' + error);
+    }
   }
 
   // mở rộng quyền: nếu có create/edit/delete thì tự động thêm view
